+++ crates/blockchain/payload.rs
@@ -2,6 +2,7 @@ use std::{
     cmp::{max, Ordering},
     collections::HashMap,
     ops::Div,
+    sync::Arc,
     time::Instant,
 };
 
@@ -316,7 +317,7 @@ impl Blockchain {
     /// Returns two transaction queues, one for plain and one for blob txs
     pub fn fetch_mempool_transactions(
         &self,
-        context: &mut PayloadBuildContext,
+        context: &PayloadBuildContext,
     ) -> Result<(TransactionQueue, TransactionQueue), ChainError> {
         let tx_filter = PendingTxFilter {
             /*TODO(https://github.com/lambdaclass/ethrex/issues/680): add tip filter */
@@ -349,6 +350,12 @@ impl Blockchain {
     /// Fills the payload with transactions taken from the mempool
     /// Returns the block value
     pub fn fill_transactions(&self, context: &mut PayloadBuildContext) -> Result<(), ChainError> {
+        // Check if we have enough gas at all
+        if context.remaining_gas < TX_GAS_COST {
+            debug!("No more gas to run transactions");
+            return Ok(());
+        };
+
         let chain_config = context.chain_config()?;
         let max_blob_number_per_block = chain_config
             .get_fork_blob_schedule(context.payload.header.timestamp)
@@ -358,6 +365,10 @@ impl Blockchain {
         debug!("Fetching transactions from mempool");
         // Fetch mempool transactions
         let (mut plain_txs, mut blob_txs) = self.fetch_mempool_transactions(context)?;
+
+        // Txs to remove from the mempool in a batch.
+        let mut transactions_to_remove = Vec::with_capacity(plain_txs.txs.len());
+
         // Execute and add transactions to payload (if suitable)
         loop {
             // Check if we have enough gas to run more transactions
@@ -405,7 +416,7 @@ impl Blockchain {
                 // Pull transaction from the mempool
                 debug!("Ignoring replay-protected transaction: {}", tx_hash);
                 txs.pop();
-                self.remove_transaction_from_pool(&head_tx.tx.compute_hash())?;
+                transactions_to_remove.push(head_tx.tx.compute_hash());
                 continue;
             }
 
@@ -419,7 +430,7 @@ impl Blockchain {
                 Ok(receipt) => {
                     txs.shift()?;
                     // Pull transaction from the mempool
-                    self.remove_transaction_from_pool(&head_tx.tx.compute_hash())?;
+                    transactions_to_remove.push(head_tx.tx.compute_hash());
 
                     metrics!(METRICS_TX.inc_tx_with_status_and_type(
                         MetricsTxStatus::Succeeded,
@@ -444,6 +455,8 @@ impl Blockchain {
             // Save receipt for hash calculation
             context.receipts.push(receipt);
         }
+
+        self.remove_transactions_from_pool(&transactions_to_remove)?;
         Ok(())
     }
 
@@ -489,7 +502,7 @@ impl Blockchain {
         let prev_blob_gas = context.payload.header.blob_gas_used.unwrap_or_default();
         context.payload.header.blob_gas_used =
             Some(prev_blob_gas + blobs_bundle.blobs.len() as u64 * GAS_PER_BLOB);
-        context.blobs_bundle += blobs_bundle;
+        context.blobs_bundle += (*blobs_bundle).clone();
         Ok(receipt)
     }
 
@@ -556,14 +569,14 @@ pub struct TransactionQueue {
     // The first transaction for each account along with its tip, sorted by highest tip
     heads: Vec<HeadTransaction>,
     // The remaining txs grouped by account and sorted by nonce
-    txs: HashMap<Address, Vec<MempoolTransaction>>,
+    txs: HashMap<Address, Vec<Arc<MempoolTransaction>>>,
     // Base Fee stored for tip calculations
     base_fee: Option<u64>,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq)]
 pub struct HeadTransaction {
-    pub tx: MempoolTransaction,
+    pub tx: Arc<MempoolTransaction>,
     tip: u64,
 }
 
@@ -584,7 +597,7 @@ impl From<HeadTransaction> for Transaction {
 impl TransactionQueue {
     /// Creates a new TransactionQueue from a set of transactions grouped by sender and sorted by nonce
     fn new(
-        mut txs: HashMap<Address, Vec<MempoolTransaction>>,
+        mut txs: HashMap<Address, Vec<Arc<MempoolTransaction>>>,
         base_fee: Option<u64>,
     ) -> Result<Self, ChainError> {
         let mut heads = Vec::new();